+++ changes/6457-dmontagu.md
@@ -1,0 +1,1 @@
+Fix bug where round-trip pickling/unpickling a `RootModel` would change the value of `__dict__`.
